Renames component to be more specific.
